// models/content-types/article.ts
import { type IContentItem, type Elements } from '@kontent-ai/delivery-sdk';
/**
 * Generated by '@kontent-ai/model-generator@5.10.0'
 *
 * Article
 * Id: d38e5c5f-c6de-4765-ab2c-c2b084507217
 * Codename: article
 */
export type Article = IContentItem<{
  /**
   * Body (rich_text)
   * Required: false
   * Id: 7b578e9c-1c4a-401d-9116-4aa06f5b4da7
   * Codename: body
   *
   * This is rich text field that you can use to add formatted text and compose larger content out of smaller pieces. You can also add images, links, and other content.
   */
  body: Elements.RichTextElement;

  /**
   * Title (text)
   * Required: false
   * Id: 1c0bb109-594e-4c15-b9cd-437387add2f6
   * Codename: title
   *
   * This is a simple text field. It defines the title for your article (and can be different than the name in Kontent.ai). Text fields don't store any formatting.
   */
  title: Elements.TextElement;

  /**
   * URL (url_slug)
   * Required: true
   * Id: 615d1f3d-5109-445f-bd2d-4bc8e53e2eb8
   * Codename: url
   *
   * This URL slug is automatically generated from the title. But you can also change it to be more SEO-friendly.
   */
  url: Elements.UrlSlugElement;
}>;